// src/features/productFilters/model/types/productFiltersSchema.ts
import type {PriceRangeType} from "@/entities/product/model/types/Product";

export interface productFiltersSchema {
    filters: {
        priceRange: PriceRangeType
        countries: string[]
        brands: string[]
        inStock: boolean;
        sortBy: SortType;
        sortOrder: OrderType;
    }
    isOpen: boolean
}

export type SortType = 'price' | 'rating' | 'name'
export type OrderType = 'asc' | 'desc'